feat: trying to get findings to return something
